Refactor BirthdroidListAdapter into own file
